+++ packages/sdk/contractkit/src/wrappers/MetaTransactionWalletDeployer.ts
@@ -4,3 +4,5 @@ import { BaseWrapper, proxySend } from './BaseWrapper'
 export class MetaTransactionWalletDeployerWrapper extends BaseWrapper<MetaTransactionWalletDeployer> {
   deploy = proxySend(this.connection, this.contract.methods.deploy)
 }
+
+export type MetaTransactionWalletDeployerWrapperType = MetaTransactionWalletDeployerWrapper