+++ packages/sdk/contractkit/src/wrappers/MultiSig.ts
@@ -119,3 +119,5 @@ export class MultiSigWrapper extends BaseWrapper<MultiSig> {
     return res
   }
 }
+
+export type MultiSigWrapperType = MultiSigWrapper