+++ packages/sdk/contractkit/src/wrappers/SortedOracles.ts
@@ -320,3 +320,5 @@ export class SortedOraclesWrapper extends BaseWrapper<SortedOracles> {
     }
   }
 }
+
+export type SortedOraclesWrapperType = SortedOraclesWrapper