+++ packages/sdk/contractkit/src/wrappers/StableTokenWrapper.ts
@@ -135,3 +135,5 @@ export class StableTokenWrapper extends CeloTokenWrapper<StableToken> {
     }
   }
 }
+
+export type StableTokenWrapperType = StableTokenWrapper